# swagger.yaml
openapi: 3.0.3
info:
  title: Bell Project API
  description: QR-based restaurant ordering system API
  version: 1.0.0
  contact:
    name: Bell Project Team
    email: support@bellproject.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.bellproject.com/dev
    description: Development server
  - url: https://api.bellproject.com/prod
    description: Production server
...
paths:
...
  /pos/print:
    post:
      tags:
        - pos
      summary: Create POS print job
      description: Submit a print job to the POS system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPosPrintRequest'
      responses:
        '201':
          description: Print job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPosPrintResponse'
        '400':
          description: Bad request - Invalid print job data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pos/print/{jobId}:
    get:
      tags:
        - pos
      summary: Get POS print job status
      description: Check the status of a print job
      parameters:
        - name: jobId
          in: path
          required: true
          description: Print job identifier
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Print job status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPosPrintResponse'
        '404':
          description: Print job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /pos/reprint:
    post:
      tags:
        - pos
      summary: Reprint POS job
      description: Reprint an existing print job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostPosReprintRequest'
      responses:
        '201':
          description: Reprint job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPosPrintResponse'
        '400':
          description: Bad request - Invalid reprint data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payment/callback/naverpay:
    post:
      tags:
        - payment
      summary: NaverPay payment callback
      description: Handle payment callback from NaverPay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCallbackRequest'
      responses:
        '200':
          description: Callback processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCallbackResponse'
        '400':
          description: Bad request - Invalid callback data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payment/callback/kakaopay:
    post:
      tags:
        - payment
      summary: KakaoPay payment callback
      description: Handle payment callback from KakaoPay
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCallbackRequest'
      responses:
        '200':
          description: Callback processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCallbackResponse'
        '400':
          description: Bad request - Invalid callback data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    PostPosPrintRequest:
      type: object
      required:
        - orderId
        - restaurantId
        - printerType
        - items
      properties:
        orderId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        restaurantId:
          type: string
          example: "rest_001"
        printerType:
          type: string
          enum: [receipt, kitchen]
          example: "receipt"
        items:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "Classic Burger"
              quantity:
                type: integer
                example: 2
              price:
                type: number
                format: float
                example: 12.99
              options:
                type: array
                items:
                  type: string
                example: ["Large", "Extra Cheese"]
    PostPosPrintResponse:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
          example: "job_550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [QUEUED, PRINTING, COMPLETED, FAILED]
          example: "QUEUED"
        createdAt:
          type: string
          format: date-time
          example: "2023-08-15T12:30:45.123Z"
    GetPosPrintResponse:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
          example: "job_550e8400-e29b-41d4-a716-446655440000"
        orderId:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [QUEUED, PRINTING, COMPLETED, FAILED]
          example: "COMPLETED"
        createdAt:
          type: string
          format: date-time
          example: "2023-08-15T12:30:45.123Z"
        completedAt:
          type: string
          format: date-time
          example: "2023-08-15T12:31:45.123Z"
    PostPosReprintRequest:
      type: object
      required:
        - originalJobId
      properties:
        originalJobId:
          type: string
          format: uuid
          example: "job_550e8400-e29b-41d4-a716-446655440000"
    PaymentCallbackRequest:
      type: object
      required:
        - orderId
        - transactionId
        - status
        - amount
      properties:
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        transactionId:
          type: string
          example: "txn_123456789"
        status:
          type: string
          enum: [SUCCESS, FAILED, CANCELLED]
          example: "SUCCESS"
        amount:
          type: number
          format: float
          example: 25.98
        timestamp:
          type: string
          format: date-time
          example: "2023-08-15T12:35:45.123Z"
    PaymentCallbackResponse:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [PROCESSED, REJECTED]
          example: "PROCESSED"
        message:
          type: string
          example: "Payment processed successfully"
    ErrorResponse:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          format: uri
          example: "https://httpstatuses.com/400"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: "The request is missing required parameters"
        instance:
          type: string
          format: uri
          example: "/menu"